4.10. Перенес функцию генерации рандомного числа в общий utils. Сделал дополнительное задание, на странице login, добавил компонент перехода на рандомный город.
