+++ src/mock/utils.ts
@@ -1,16 +1,4 @@
-const generateRandomNumber = (min: number = 0, max: number = 1000, isInteger: boolean = true): number => {
-
-  let processedMin: number = Math.min(min, max);
-  let processedMax: number = Math.max(min, max);
-
-  if (isInteger) {
-    processedMin = Math.ceil(processedMin);
-    processedMax = Math.floor(processedMax);
-    return Math.floor(Math.random() * (processedMax - processedMin + 1) + processedMin);
-  }
-
-  return Math.round((Math.random() * (processedMax - processedMin) + processedMin) * 10) / 10;
-};
+import { generateRandomNumber } from '../utils';
 
 const flipCoin = (): boolean => Boolean(Math.round(Math.random()));
 
@@ -26,4 +14,4 @@ const getUniqRandomElement = <T>(items: T[]): (() => T) => (
   }
 );
 
-export{ generateRandomNumber, flipCoin, getRandomElement, getUniqRandomElement };
+export{ flipCoin, getRandomElement, getUniqRandomElement };
