+++ src/utils.ts
@@ -2,6 +2,20 @@ import { CITIES, AppRoute } from './consts';
 import { Offer, Cities, Page } from './types';
 
 
+const generateRandomNumber = (min: number = 0, max: number = 1000, isInteger: boolean = true): number => {
+
+  let processedMin: number = Math.min(min, max);
+  let processedMax: number = Math.max(min, max);
+
+  if (isInteger) {
+    processedMin = Math.ceil(processedMin);
+    processedMax = Math.floor(processedMax);
+    return Math.floor(Math.random() * (processedMax - processedMin + 1) + processedMin);
+  }
+
+  return Math.round((Math.random() * (processedMax - processedMin) + processedMin) * 10) / 10;
+};
+
 const capitalizeFirstLetter = (inputWord: string): string => inputWord[0].toUpperCase() + inputWord.slice(1);
 
 const getPageName = (path: string): Page => {
@@ -27,4 +41,4 @@ const sortOffersByCity = (offers: Offer[]) =>
   );
 
 
-export { capitalizeFirstLetter, getPageName, sortOffersByCity };
+export { generateRandomNumber, capitalizeFirstLetter, getPageName, sortOffersByCity };
